chore: remove lldb related thing from main.rs
